requirements.txt
cffi==1.17.1
cryptography==45.0.6
greenlet==3.2.3
loguru==0.7.3
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
